Renaming packages, some UI fixes, adding jvmStatic
