--- library/src/enabled/java/com/gmerino/tweak/di/TweaksModule.kt
@@ -1,13 +1,0 @@
-package com.gmerino.tweak.di
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-
-@Module
-class TweaksModule(private val context: Context) {
-
-    @Provides
-    fun provideContext() = context
-
-}
