+++ library/src/enabled/java/com/gmerinojimenez/tweak/di/TweaksModule.kt
@@ -1,0 +1,13 @@
+package com.gmerinojimenez.tweak.di
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+
+@Module
+class TweaksModule(private val context: Context) {
+
+    @Provides
+    fun provideContext() = context
+
+}
